requirements-azure.txt
Django==5.2.9
psycopg2-binary==2.9.9
django-environ==0.11.2
pandas==2.1.4
plotly==5.17.0
openpyxl==3.1.2
xlrd==2.0.1
reportlab==4.0.8
whitenoise==6.6.0
gunicorn==21.2.0
openai==1.3.0
djangorestframework==3.14.0